+++ docx4j-core/src/main/java/org/docx4j/fonts/fop/fonts/truetype/OFFontLoader.java
@@ -200,7 +200,7 @@ public class OFFontLoader extends FontLoader {
                 multiFont.setCIDType(CIDFontType.CIDTYPE2);
             }
             multiFont.setWidthArray(otf.getWidths());
-            multiFont.setBBoxArray(otf.getBoundingBoxes());
+            multiFont.setBBoxArray(otf.getBoundingBoxesPacked());
         } else {
             singleFont.setFontType(FontType.TRUETYPE);
             singleFont.setEncoding(otf.getCharSetName());